# api/main.py
"""Read-only FastAPI service for ForroVivo Linguistic Research datasets."""

from __future__ import annotations

import hashlib
import re
import secrets
from datetime import datetime, timezone
from typing import Any, Dict, Optional

from fastapi import FastAPI, Query, Request
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import FileResponse, JSONResponse, RedirectResponse
from uvicorn.middleware.proxy_headers import ProxyHeadersMiddleware

from api import REPO_ROOT
from api.settings import (
    API_FAMILY,
    API_HOST,
    API_ORIGIN,
    APP_STORE_URL,
    CORS_ALLOW_ORIGIN,
    CORS_EXPOSE_HEADERS,
    GITHUB_URL,
    LICENSE_URL,
    PROJECT_LICENSE,
    SITE_ORIGIN,
)
from api.knowledge import (
    DOCUMENT_COLLECTIONS,
    collection_directory,
    knowledge_base_map,
    languages_catalog,
    page_rows,
    record_by_id,
    search_collection,
    search_dataset,
)
from api.catalog import FAMILY_INDEXES
from api.store import (
    Store,
    audio_file,
    lookup_headword,
    present_entry,
    search_entries,
)

STORE = Store.load()
FAMILIES = frozenset(family for family, _ in FAMILY_INDEXES)
DEFAULT_LIMIT = 50
MAX_LIMIT = 1000

app = FastAPI(
    title="ForroVivo Linguistic Research API",
    description=(
        "ForroVivo is the platform and ecosystem. This API is the machine-readable "
        "layer of the Linguistic Research initiative within it. "
        "Public host: https://api.forrovivo.com. "
        "Each dataset is isolated. Missing terms return TERM_NOT_FOUND. "
        "This API does not invent translations or merge languages."
    ),
    version="2.4.0",
    servers=[
        {"url": API_ORIGIN, "description": "Production"},
        {"url": "http://127.0.0.1:8000", "description": "Local"},
    ],
)
app.add_middleware(ProxyHeadersMiddleware, trusted_hosts="*")
app.add_middleware(
    CORSMiddleware,
    allow_origins=[CORS_ALLOW_ORIGIN],
    allow_methods=["GET", "HEAD", "POST", "OPTIONS"],
    allow_headers=["Accept", "Content-Type", "If-None-Match", "Authorization", "X-Api-Key"],
    expose_headers=CORS_EXPOSE_HEADERS,
    max_age=86400,
)

KEY_PREFIX = "fv_live_"
EMAIL_PATTERN = re.compile(r"^[^\s@]+@[^\s@]+\.[^\s@]+$")
KEY_BY_HASH: Dict[str, Dict[str, str]] = {}
HASH_BY_EMAIL: Dict[str, str] = {}


def _sha256_hex(value: str) -> str:
    return hashlib.sha256(value.encode("utf-8")).hexdigest()


@app.middleware("http")
async def developer_headers(request: Request, call_next):
    response = await call_next(request)
    response.headers["API-Version"] = API_FAMILY
    response.headers["Link"] = (
        f'<{API_ORIGIN}/v1>; rel="index", '
        f'<{API_ORIGIN}/v1/openapi.yaml>; rel="describedby"; type="application/yaml", '
        f'<{GITHUB_URL}>; rel="source", '
        f'<{LICENSE_URL}>; rel="license"'
    )
    return response


def request_origin(request: Request) -> str:
    return str(request.base_url).rstrip("/")


def present(dataset, entry, request: Request) -> Dict[str, Any]:
    return present_entry(dataset, entry, base_url=request_origin(request))


def term_not_found(dataset) -> JSONResponse:
    return JSONResponse(status_code=404, content=dataset.missing_term)


def dataset_not_found(key: str) -> JSONResponse:
    return JSONResponse(
        status_code=404,
        content={
            "status": "error",
            "code": "DATASET_NOT_FOUND",
            "message": "No isolated dataset is published at this path.",
            "dataset": key,
        },
    )


def resolve(family: str, variety: Optional[str] = None):
    if family not in FAMILIES:
        return None, dataset_not_found(family if not variety else f"{family}/{variety}")
    key = STORE.resolve_key(family, variety)
    dataset = STORE.get(key)
    if dataset is None:
        return None, dataset_not_found(key)
    return dataset, None


def envelope(dataset, extra: Dict[str, Any]) -> Dict[str, Any]:
    license_payload = dataset.document.get("license") or PROJECT_LICENSE
    folder = dataset.ref.json_path.parent.relative_to(REPO_ROOT).as_posix()
    payload = {
        "dataset": dataset.ref.key,
        "kind": dataset.kind,
        "language": dataset.document.get("language"),
        "attribution": {
            "dataset": dataset.ref.key,
            "github": f"{GITHUB_URL}/tree/main/{folder}",
            "sources": f"/v1/{dataset.ref.key}/sources",
            "license": license_payload,
            "license_url": LICENSE_URL,
        },
    }
    if dataset.ref.canonical_key:
        payload["canonical_dataset"] = dataset.ref.canonical_key
    payload.update(extra)
    return payload


def dataset_required(collection: str) -> JSONResponse:
    return JSONResponse(
        status_code=400,
        content={
            "status": "error",
            "code": "DATASET_REQUIRED",
            "message": "This collection is isolated. Name one dataset; do not search across languages.",
            "collection": collection,
            "example": f"/v1/saotome/forro/{collection}" if collection != "search" else "/v1/search?dataset=saotome/forro&q=kume",
        },
    )


def present_hit(dataset, hit: Dict[str, Any], request: Request) -> Dict[str, Any]:
    if hit.get("collection") == "entries":
        return {
            "collection": "entries",
            "item": present(dataset, hit["item"], request),
        }
    return hit


def present_records(dataset, collection: str, rows: list, request: Request) -> list:
    if collection == "entries":
        return [present(dataset, row, request) for row in rows]
    return rows


@app.get("/health")
def health() -> Dict[str, str]:
    return {"status": "ok"}


@app.get("/v1/health")
def versioned_health() -> Dict[str, str]:
    return {"status": "ok", "api": API_FAMILY}


@app.get("/")
def site_root() -> RedirectResponse:
    return RedirectResponse(url="/v1", status_code=307)


@app.post("/v1/keys")
def issue_api_key(payload: Dict[str, Any]) -> Any:
    email = payload.get("email") if isinstance(payload, dict) else None
    normalized = email.strip().lower() if isinstance(email, str) else ""
    if not normalized or EMAIL_PATTERN.match(normalized) is None:
        return JSONResponse(
            {
                "status": "error",
                "code": "INVALID_EMAIL",
                "message": "A valid email address is required.",
            },
            status_code=400,
        )
    presented = KEY_PREFIX + secrets.token_hex(24)
    digest = _sha256_hex(presented)
    previous = HASH_BY_EMAIL.get(normalized)
    if previous:
        KEY_BY_HASH.pop(previous, None)
    record = {
        "email": normalized,
        "created_at": datetime.now(timezone.utc).isoformat(),
        "prefix": presented[:16],
    }
    KEY_BY_HASH[digest] = record
    HASH_BY_EMAIL[normalized] = digest
    return JSONResponse(
        {
            "status": "ok",
            "key": presented,
            "email": normalized,
            "prefix": record["prefix"],
            "created_at": record["created_at"],
            "shown_once": True,
            "usage": "Authorization: Bearer <key>",
        },
        status_code=201,
    )


@app.get("/v1")
def root() -> Dict[str, Any]:
    return {
        "name": "ForroVivo Linguistic Research API",
        "api": API_FAMILY,
        "version": "2.4.0",
        "platform": "ForroVivo",
        "initiative": "Linguistic Research",
        "founder": "Henriques Pontes",
        "idealist": "Henriques Pontes",
        "cofounders": ["Luis Lima"],
        "host": API_HOST,
        "url": API_ORIGIN,
        "homepage": SITE_ORIGIN,
        "github": GITHUB_URL,
        "app_store": APP_STORE_URL,
        "project_start_date": "2023-03-23",
        "authentication": (
            "Optional. Public GET stays open. Issue a key with POST /v1/keys. "
            "Send Authorization: Bearer <key>."
        ),
        "naming": "/v1/{family}/{variety}/{collection}",
        "cors": "Any origin. GET, HEAD, and OPTIONS. Credentials are not used.",
        "rate_limit": "Fair-use per client. See RateLimit-Policy and Retry-After.",
        "attribution": (
            "Each lexicon keeps its cited sources. Project materials are CC BY 4.0. "
            "Source extracts keep their original terms. See Link: rel=license and rel=source."
        ),
        "principle": "Zero hallucination. Missing data is preferable to incorrect data.",
        "isolation": (
            "Each path serves one dataset. Parent indexes are not merged lexicons. "
            "data/angola_dataset/ is an Angola country index (Contruy, Umbundu, Kimbundu, Kikongo). "
            "It is not Angolar / Ngola. Other country folders are indexes of their languages."
        ),
        "graph": (
            "Each entry includes an attested relation graph: "
            "means (Portuguese / English concepts), belongs_to (one language), "
            "related_to (grammar, culture), appears_in (proverb, story), "
            "documented_by (source). Missing edges stay empty. Edges never cross folders."
        ),
        "license": PROJECT_LICENSE,
        "docs": f"{API_ORIGIN}/docs",
        "openapi": f"{API_ORIGIN}/v1/openapi.yaml",
        "catalog": f"{API_ORIGIN}/v1/datasets",
        "knowledge_base": f"{API_ORIGIN}/v1/kb",
        "languages": f"{API_ORIGIN}/v1/languages",
    }


@app.get("/v1/datasets")
def list_datasets() -> Dict[str, Any]:
    return {"datasets": STORE.catalog()}


@app.get("/v1/kb")
def knowledge_base() -> Dict[str, Any]:
    return knowledge_base_map(API_ORIGIN)


@app.get("/v1/languages")
def list_languages() -> Dict[str, Any]:
    return languages_catalog(STORE, API_ORIGIN)


@app.get("/v1/search")
def search_isolated(
    request: Request,
    dataset: Optional[str] = Query(None, description="Isolated dataset key, for example saotome/forro."),
    q: Optional[str] = Query(None, description="Search inside that dataset only."),
    offset: int = Query(0, ge=0),
    limit: int = Query(DEFAULT_LIMIT, ge=1, le=MAX_LIMIT),
) -> Any:
    if not dataset:
        return dataset_required("search")
    parts = dataset.strip("/").split("/")
    family = parts[0]
    variety = parts[1] if len(parts) > 1 else None
    loaded, error = resolve(family, variety)
    if error is not None:
        return error
    if loaded.kind == "index":
        return term_not_found(loaded)
    if not q:
        return JSONResponse(
            status_code=400,
            content={
                "status": "error",
                "code": "QUERY_REQUIRED",
                "message": "Search requires q= inside one isolated dataset.",
                "dataset": loaded.ref.key,
            },
        )
    rows = search_dataset(loaded, q)
    page, total = page_rows(rows, offset, limit)
    return envelope(
        loaded,
        {
            "collection": "search",
            "query": q,
            "total": total,
            "offset": offset,
            "limit": limit,
            "results": [present_hit(loaded, hit, request) for hit in page],
        },
    )


@app.get("/v1/entries")
@app.get("/v1/grammar")
@app.get("/v1/expressions")
@app.get("/v1/proverbs")
@app.get("/v1/culture")
@app.get("/v1/food")
@app.get("/v1/music")
@app.get("/v1/dance")
@app.get("/v1/folklore")
@app.get("/v1/stories")
@app.get("/v1/places")
@app.get("/v1/sources")
def knowledge_collection_index(request: Request) -> Dict[str, Any]:
    collection = request.url.path.rstrip("/").rsplit("/", 1)[-1]
    return collection_directory(STORE, collection)


@app.get("/v1/{family}/lookup")
@app.get("/v1/{family}/{variety}/lookup")
def lookup(
    request: Request,
    family: str,
    variety: Optional[str] = None,
    headword: str = Query(..., min_length=1),
) -> Any:
    dataset, error = resolve(family, variety)
    if error is not None:
        return error
    if dataset.kind == "index":
        return term_not_found(dataset)
    match, entries = lookup_headword(dataset, headword)
    if not entries:
        return term_not_found(dataset)
    return envelope(
        dataset,
        {
            "query": {"headword": headword},
            "match": match,
            "count": len(entries),
            "entries": [present(dataset, entry, request) for entry in entries],
        },
    )


@app.get("/v1/{family}/entries/{entry_id}")
@app.get("/v1/{family}/{variety}/entries/{entry_id}")
def get_entry(request: Request, entry_id: str, family: str, variety: Optional[str] = None) -> Any:
    dataset, error = resolve(family, variety)
    if error is not None:
        return error
    if dataset.kind == "index":
        return term_not_found(dataset)
    entry = dataset.by_id.get(entry_id)
    if entry is None:
        return term_not_found(dataset)
    return envelope(dataset, {"entry": present(dataset, entry, request)})


@app.get("/v1/{family}/entries")
@app.get("/v1/{family}/{variety}/entries")
def list_entries(
    request: Request,
    family: str,
    variety: Optional[str] = None,
    q: Optional[str] = Query(None, description="Search inside this dataset only."),
    offset: int = Query(0, ge=0),
    limit: int = Query(DEFAULT_LIMIT, ge=1, le=MAX_LIMIT),
) -> Any:
    dataset, error = resolve(family, variety)
    if error is not None:
        return error
    if dataset.kind == "index":
        return term_not_found(dataset)
    rows = search_entries(dataset, q) if q else dataset.entries
    page = rows[offset : offset + limit]
    return envelope(
        dataset,
        {
            "query": q,
            "total": len(rows),
            "offset": offset,
            "limit": limit,
            "entries": [present(dataset, entry, request) for entry in page],
        },
    )


@app.get("/v1/{family}/audio/{filename}")
@app.get("/v1/{family}/{variety}/audio/{filename}")
def get_audio(filename: str, family: str, variety: Optional[str] = None) -> Any:
    dataset, error = resolve(family, variety)
    if error is not None:
        return error
    served = dataset
    if dataset.ref.canonical_key:
        served = STORE.require(dataset.ref.canonical_key)
    path = audio_file(served, filename)
    if path is None:
        return JSONResponse(
            status_code=404,
            content={
                "status": "error",
                "code": "AUDIO_NOT_FOUND",
                "message": "No audio file with that name is stored in this dataset.",
                "dataset": served.ref.key,
            },
        )
    return FileResponse(path, media_type="audio/mpeg")


def list_knowledge_collection(
    request: Request,
    family: str,
    variety: Optional[str] = None,
    q: Optional[str] = Query(None),
    offset: int = Query(0, ge=0),
    limit: int = Query(DEFAULT_LIMIT, ge=1, le=MAX_LIMIT),
) -> Any:
    collection = request.url.path.rstrip("/").rsplit("/", 1)[-1]
    dataset, error = resolve(family, variety)
    if error is not None:
        return error
    if dataset.kind == "index":
        return term_not_found(dataset)
    rows = search_collection(dataset, collection, q)
    page, total = page_rows(rows, offset, limit)
    return envelope(
        dataset,
        {
            "collection": collection,
            "query": q,
            "total": total,
            "offset": offset,
            "limit": limit,
            "items": present_records(dataset, collection, page, request),
        },
    )


def get_knowledge_item(
    request: Request,
    item_id: str,
    family: str,
    variety: Optional[str] = None,
) -> Any:
    collection = request.url.path.rstrip("/").rsplit("/", 2)[-2]
    dataset, error = resolve(family, variety)
    if error is not None:
        return error
    if dataset.kind == "index":
        return term_not_found(dataset)
    item = record_by_id(dataset, collection, item_id)
    if item is None:
        return term_not_found(dataset)
    payload = item
    if collection == "entries":
        payload = present(dataset, item, request)
    return envelope(dataset, {"collection": collection, "item": payload})


def search_one_dataset(
    request: Request,
    family: str,
    variety: Optional[str] = None,
    q: Optional[str] = Query(None, description="Search inside this dataset only."),
    offset: int = Query(0, ge=0),
    limit: int = Query(DEFAULT_LIMIT, ge=1, le=MAX_LIMIT),
) -> Any:
    dataset, error = resolve(family, variety)
    if error is not None:
        return error
    if dataset.kind == "index":
        return term_not_found(dataset)
    if not q:
        return JSONResponse(
            status_code=400,
            content={
                "status": "error",
                "code": "QUERY_REQUIRED",
                "message": "Search requires q= inside one isolated dataset.",
                "dataset": dataset.ref.key,
            },
        )
    rows = search_dataset(dataset, q)
    page, total = page_rows(rows, offset, limit)
    return envelope(
        dataset,
        {
            "collection": "search",
            "query": q,
            "total": total,
            "offset": offset,
            "limit": limit,
            "results": [present_hit(dataset, hit, request) for hit in page],
        },
    )


for _name in (*DOCUMENT_COLLECTIONS, "sources"):
    app.add_api_route(f"/v1/{{family}}/{_name}", list_knowledge_collection, methods=["GET"])
    app.add_api_route(f"/v1/{{family}}/{{variety}}/{_name}", list_knowledge_collection, methods=["GET"])
    app.add_api_route(f"/v1/{{family}}/{_name}/{{item_id}}", get_knowledge_item, methods=["GET"])
    app.add_api_route(f"/v1/{{family}}/{{variety}}/{_name}/{{item_id}}", get_knowledge_item, methods=["GET"])

app.add_api_route("/v1/{family}/search", search_one_dataset, methods=["GET"])
app.add_api_route("/v1/{family}/{variety}/search", search_one_dataset, methods=["GET"])


@app.get("/v1/{family}")
@app.get("/v1/{family}/{variety}")
def get_dataset(family: str, variety: Optional[str] = None) -> Any:
    dataset, error = resolve(family, variety)
    if error is not None:
        return error
    return dataset.metadata()