# api/openapi.yaml
openapi: 3.0.3
info:
  title: ForroVivo Linguistic Research API
  version: "2.4.0"
  description: |
    Public, read-only linguistic API. Public GET needs no key.

    **Versioning.** All data routes live under `/v1`. `/` redirects to `/v1`.
    Responses send `API-Version: v1`. The JSON field `api` is the URL family;
    `version` is the release.

    **Naming.** Lexicons use `/v1/{family}/{variety}/{collection}`.
    Country indexes are `/v1/{family}` and are not merged dictionaries.

    **Authentication.** Optional. Issue a key with `POST /v1/keys`. Send
    `Authorization: Bearer <key>`. Public lookup stays open without a key.

    **CORS.** Any origin. Credentials are not used.

    **Rate limits.** Fair-use per client. Exceeding the policy returns `429`
    `RATE_LIMITED` with `Retry-After`. Read `RateLimit-Policy`.

    **Attribution.** Each entry keeps its cited source. Lookup envelopes include
    `attribution`. Responses send `Link` with `rel=source` (GitHub) and
    `rel=license`. Project materials are CC BY 4.0; source extracts keep their
    original terms. Gloss fields are concepts the headword means, not proof the
    word belongs to Portuguese or English.

    Isolation: missing terms return TERM_NOT_FOUND. Search never crosses folders.
    Knowledge collections stay empty until a cited source fills them.
  license:
    name: Mixed. Project materials CC BY 4.0. Source extracts keep original terms.
    url: https://creativecommons.org/licenses/by/4.0/
  contact:
    name: ForroVivo Linguistic Research
    url: https://github.com/Forrovivo/linguistic-research
servers:
  - url: https://api.forrovivo.com
    description: Production
  - url: http://127.0.0.1:8787
    description: Cloudflare Worker local
  - url: http://127.0.0.1:8000
    description: Python local
tags:
  - name: Catalog
  - name: Lookup
  - name: Knowledge
components:
  headers:
    ApiVersion:
      schema:
        type: string
        example: v1
      description: URL family. Currently v1.
    Link:
      schema:
        type: string
      description: index, describedby (OpenAPI), source (GitHub), and license.
    RateLimitPolicy:
      schema:
        type: string
      description: Fair-use policy for this deployment.
    RetryAfter:
      schema:
        type: string
      description: Seconds to wait after RATE_LIMITED.
  responses:
    RateLimited:
      description: RATE_LIMITED
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
        RateLimit-Policy:
          $ref: "#/components/headers/RateLimitPolicy"
        API-Version:
          $ref: "#/components/headers/ApiVersion"
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: error
              code:
                type: string
                example: RATE_LIMITED
              message:
                type: string
paths:
  /health:
    get:
      summary: Process health
      responses:
        "200":
          description: OK
  /v1/health:
    get:
      tags: [Catalog]
      summary: Versioned health
      responses:
        "200":
          description: OK
  /v1/openapi.yaml:
    get:
      tags: [Catalog]
      summary: OpenAPI contract
      responses:
        "200":
          description: application/yaml
  /v1:
    get:
      summary: API metadata, versioning, CORS, rate limit, and attribution
      tags: [Catalog]
      responses:
        "200":
          description: Host, license, naming, and catalog links
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/keys:
    post:
      summary: Issue or replace an optional API key
      tags: [Catalog]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
      responses:
        "201":
          description: Key shown once
        "400":
          description: INVALID_EMAIL
  /v1/datasets:
    get:
      summary: Isolated dataset catalog
      responses:
        "200":
          description: One object per dataset; indexes are not merged lexicons
  /v1/kb:
    get:
      summary: Knowledge Base map
      responses:
        "200":
          description: Isolated collection list (languages, entries, grammar, search, …)
  /v1/languages:
    get:
      summary: Isolated language catalog
      responses:
        "200":
          description: One lexicon per object; parent indexes are omitted
  /v1/search:
    get:
      summary: Search one dataset
      parameters:
        - in: query
          name: dataset
          required: true
          schema:
            type: string
          description: Isolated key such as saotome/forro
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Hits from that dataset only
        "400":
          description: DATASET_REQUIRED or QUERY_REQUIRED
  /v1/grammar:
    get:
      summary: Grammar index (counts per isolated dataset; no merged records)
      responses:
        "200":
          description: Directory of /v1/{dataset}/grammar paths
  /v1/{family}/{variety}/grammar:
    get:
      summary: Grammar records in one isolated dataset
      parameters:
        - in: path
          name: family
          required: true
          schema:
            type: string
        - in: path
          name: variety
          required: true
          schema:
            type: string
        - in: query
          name: q
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Attested grammar records, or an empty list
        "404":
          description: TERM_NOT_FOUND for parent indexes
  /v1/{family}/{variety}/search:
    get:
      summary: Search entries, knowledge, and sources inside one dataset
      parameters:
        - in: path
          name: family
          required: true
          schema:
            type: string
        - in: path
          name: variety
          required: true
          schema:
            type: string
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Isolated hits
        "400":
          description: QUERY_REQUIRED
  /v1/{family}:
    get:
      summary: Dataset metadata
      parameters:
        - in: path
          name: family
          required: true
          schema:
            type: string
            enum:
              [
                saotome,
                caboverde,
                guinebissau,
                angola,
                seychelles,
                equatorialguinea,
                southafrica,
                ghana,
                sierraleone,
                rdcongo,
                mauritius,
                nigeria,
                reunion,
                centralafrican,
                cameroon,
              ]
      responses:
        "200":
          description: Metadata without the full entry list
        "404":
          description: DATASET_NOT_FOUND
  /v1/{family}/lookup:
    get:
      summary: Headword lookup in one dataset
      parameters:
        - in: path
          name: family
          required: true
          schema:
            type: string
        - in: query
          name: headword
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching entries from that dataset only
        "404":
          description: TERM_NOT_FOUND or DATASET_NOT_FOUND
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/{family}/{variety}:
    get:
      summary: Island or region dataset metadata
      parameters:
        - in: path
          name: family
          required: true
          schema:
            type: string
            enum:
              [
                saotome,
                caboverde,
                guinebissau,
                angola,
                seychelles,
                equatorialguinea,
                southafrica,
                ghana,
                sierraleone,
                rdcongo,
                mauritius,
                nigeria,
                reunion,
                centralafrican,
                cameroon,
              ]
        - in: path
          name: variety
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Metadata
        "404":
          description: DATASET_NOT_FOUND
  /v1/{family}/{variety}/lookup:
    get:
      summary: Headword lookup in one island or region
      parameters:
        - in: path
          name: family
          required: true
          schema:
            type: string
        - in: path
          name: variety
          required: true
          schema:
            type: string
        - in: query
          name: headword
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching entries from that variety only
        "404":
          description: TERM_NOT_FOUND or DATASET_NOT_FOUND
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/{family}/entries:
    get:
      summary: Paginated list or search inside one dataset
      parameters:
        - in: path
          name: family
          required: true
          schema:
            type: string
        - in: query
          name: q
          required: false
          schema:
            type: string
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Page of entries
        "404":
          description: TERM_NOT_FOUND for indexes
  /v1/{family}/entries/{entry_id}:
    get:
      summary: One entry by id
      parameters:
        - in: path
          name: family
          required: true
          schema:
            type: string
        - in: path
          name: entry_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Entry
        "404":
          description: TERM_NOT_FOUND
  /v1/{family}/audio/{filename}:
    get:
      summary: Audio file from that dataset only
      parameters:
        - in: path
          name: family
          required: true
          schema:
            type: string
        - in: path
          name: filename
          required: true
          schema:
            type: string
      responses:
        "200":
          description: audio/mpeg
        "404":
          description: AUDIO_NOT_FOUND
  /v1/{family}/{variety}/entries:
    get:
      tags: [Lookup]
      summary: Paginated entries in one island, region, or language folder
      parameters:
        - in: path
          name: family
          required: true
          schema:
            type: string
        - in: path
          name: variety
          required: true
          schema:
            type: string
        - in: query
          name: q
          required: false
          schema:
            type: string
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Page of entries from that folder only
        "404":
          description: TERM_NOT_FOUND for indexes
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/{family}/{variety}/sources:
    get:
      tags: [Knowledge]
      summary: Bibliography for one isolated folder
      parameters:
        - in: path
          name: family
          required: true
          schema:
            type: string
        - in: path
          name: variety
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cited sources for that folder
        "404":
          description: DATASET_NOT_FOUND
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/{family}/{variety}/audio/{filename}:
    get:
      tags: [Lookup]
      summary: Audio file from that folder only
      parameters:
        - in: path
          name: family
          required: true
          schema:
            type: string
        - in: path
          name: variety
          required: true
          schema:
            type: string
        - in: path
          name: filename
          required: true
          schema:
            type: string
      responses:
        "200":
          description: audio/mpeg
        "404":
          description: AUDIO_NOT_FOUND